requirements.txt
pillow == 8.1.0
numpy == 1.20.1